Implemented Damage Calculation
The damage is now calculated automatically to account for resistances, vulnerabilities and immunities. In addition, the bypass rule is checked with the properties of the weapon to, well, bypass the damage reduction.

This "fix" was made relatively quickly and dirty, so revision may be required to be more in-line with the existing code.

Possible additions to the calculation could be a check to disable the automatic damage calculation while holding a specified button while applying the damage.